requirements.txt
alembic==1.4.2
black==19.10b0
Flask==1.1.1
flask-marshmallow==0.11.0
Flask-Migrate==2.5.2
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.4.1
marshmallow==3.5.1
marshmallow-sqlalchemy==0.22.2
psycopg2==2.8.4
python-dateutil==2.8.1
python-dotenv==0.11.0
python-editor==1.0.4
SQLAlchemy==1.3.15